refactor: move things from `common` to `code-cell`
Why?
It's not easy to maintain two library at the same time. Since only
`code-cell` handles main logic in our design, it's not necessary to
place these helpers in `common`.
May the cohesion be with you.
