+++ axon-ckb-contracts/common/Cargo.toml
@@ -7,10 +7,6 @@ edition = "2018"
 
 [dependencies]
 ckb-std = {version = "0.7.4",default-features = false }
-blake2b-ref = "0.3"
-#ckb-standalone-types = {version = "0.0.1-pre.1", default-features = false}
-#molecule = { version = "0.6", default-features = false }
-bit-vec = { version = "0.6", default-features = false }
 
 common-raw = { path = "../common-raw" }
