+++ axon-ckb-contracts/common/src/lib.rs
@@ -1,84 +1,57 @@
 #![no_std]
-#![feature(lang_items)]
-#![feature(alloc_error_handler)]
-#![feature(panic_info_message)]
 
 use ckb_std::ckb_constants::Source;
-
-use crate::error::CommonError;
-use bit_vec::*;
-use ckb_std::error::SysError;
-use ckb_std::high_level::{load_cell, QueryIter};
-
-pub mod cell;
-pub mod error;
-pub mod hash;
-pub mod pattern;
-
-pub const SUDT_CODEHASH: [u8; 32] = [
-    220, 119, 113, 76, 154, 36, 54, 83, 120, 144, 192, 65, 69, 88, 144, 235, 134, 189, 90, 138, 42, 237, 82, 25, 243, 92, 250, 145, 25, 53,
-    95, 95,
-];
-pub const SUDT_HASHTYPE: u8 = 0u8;
-pub const SUDT_MUSE_ARGS: &[u8] = &[];
-
-pub const EMPTY_BIT_MAP: [u8; 32] = [0; 32];
+use ckb_std::high_level::{load_cell_data, load_cell_type, load_cell_type_hash};
+use common_raw::{cell::global_config::GlobalConfigCellData, FromRaw};
 
 pub const GLOBAL_CONFIG_TYPE_HASH: [u8; 32] = [
-    245, 208, 14, 171, 59, 152, 16, 213, 201, 255, 131, 77, 22, 90, 198, 197, 156, 60, 77, 233, 241, 191, 127, 194, 187, 229, 123, 127,
-    218, 254, 226, 255,
+    176, 130, 158, 213, 216, 232, 219, 162, 15, 11, 163, 122, 141, 76, 148, 76, 101, 99, 218, 109, 18, 206, 47, 118, 31, 150, 20, 57, 223,
+    195, 32, 204,
 ];
 
-pub fn check_args_len(expected: usize, actual: usize) -> Result<(), SysError> {
-    if actual != expected {
-        return Err(SysError::Encoding);
+pub fn check_global_cell() -> Option<GlobalConfigCellData> {
+    let global_config_data = (0..)
+        .find_map(|i| {
+            let type_hash = match load_cell_type_hash(i, Source::CellDep) {
+                Ok(hash) => hash,
+                Err(err) => return Some(Err(err)),
+            }?;
+            if type_hash == GLOBAL_CONFIG_TYPE_HASH {
+                return load_cell_data(i, Source::CellDep).ok().map(|data| Ok(data));
+            }
+            None
+        })?
+        .ok()?;
+
+    let global_config_data = GlobalConfigCellData::from_raw(&global_config_data)?;
+
+    Some(global_config_data)
+}
+
+fn check_type_script(index: usize, source: Source, code_hash: &[u8], hash_type: u8) -> Option<()> {
+    let script = load_cell_type(index, source).ok()??;
+
+    if script.as_reader().code_hash().raw_data() != code_hash {
+        return None;
     }
-    Ok(())
-}
-
-pub fn get_input_cell_count() -> usize {
-    QueryIter::new(load_cell, Source::Input).count()
-}
-
-pub fn get_output_cell_count() -> usize {
-    QueryIter::new(load_cell, Source::Output).count()
-}
-
-pub fn bit_map_add(input: &[u8; 32], checker_id: u8) -> Result<[u8; 32], CommonError> {
-    let mut input = BitVec::from_bytes(&input[..]);
-
-    //should be false
-    if input.get(checker_id as usize).ok_or(CommonError::BitOperator)? {
-        return Err(CommonError::BitOperator);
-    }
-
-    input.set(checker_id as usize, true);
-
-    let mut ret = [0u8; 32];
-    ret.copy_from_slice(input.to_bytes().as_slice());
-
-    Ok(ret)
-}
-
-pub fn bit_map_remove(input: [u8; 32], checker_id: u8) -> Result<[u8; 32], CommonError> {
-    let mut input = BitVec::from_bytes(&input[..]);
-
-    //should be true
-    if !input.get(checker_id as usize).ok_or(CommonError::BitOperator)? {
-        return Err(CommonError::BitOperator);
+    if script.as_reader().hash_type().as_slice()[0] != hash_type {
+        return None;
     }
 
-    input.set(checker_id as usize, false);
-
-    let mut ret = [0u8; 32];
-    ret.copy_from_slice(&input.to_bytes().as_slice()[0..32]);
-
-    Ok(ret)
+    Some(())
 }
 
-//check if given number is bit-marked in input array
-pub fn bit_map_marked(input: [u8; 32], checker_id: u8) -> Result<bool, CommonError> {
-    let input = BitVec::from_bytes(&input[..]);
+pub fn check_code_cell() -> Option<()> {
+    /*
+    CollatorUnlockBond,
+    Dep:    0 Global Config Cell
+    Dep:    1 .....
+    Code Cell                   ->          Code Cell
+    ...
+    */
 
-    Ok(input.get(checker_id as usize).ok_or(CommonError::BitOperator)?)
+    let global = check_global_cell()?;
+
+    check_type_script(0, Source::Input, &global.code_cell_type_codehash, global.code_cell_type_hashtype)?;
+    check_type_script(0, Source::Output, &global.code_cell_type_codehash, global.code_cell_type_hashtype)
 }
