+++ axon-ckb-contracts/contracts/always-success/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2018"
 
 [dependencies]
 ckb-std = {version = "0.7.4",default-features = false }
-common = {path = "../../common"}
 
 [features]
 default = ["ckb-std/ckb-types","ckb-std/allocator"]
