+++ axon-ckb-contracts/contracts/code-cell/src/checker_quit_sidechain.rs
@@ -1,17 +1,17 @@
 use ckb_std::ckb_constants::Source;
 
-use common::bit_map_remove;
 use common_raw::{
     cell::{
         checker_bond::{CheckerBondCellData, CheckerBondCellLockArgs},
         checker_info::{CheckerInfoCellData, CheckerInfoCellMode},
+        code::CodeCellData,
         sidechain_config::SidechainConfigCellData,
     },
     witness::checker_quit_sidechain::CheckerQuitSidechainWitness,
     FromRaw,
 };
 
-use crate::{common::*, error::Error};
+use crate::{cell::*, common::*, error::Error};
 
 const CONFIG_INPUT: CellOrigin = CellOrigin(1, Source::Input);
 const CHECKER_BOND_INPUT: CellOrigin = CellOrigin(2, Source::Input);
@@ -32,6 +32,8 @@ pub fn checker_quit_sidechain(raw_witness: &[u8], signer: [u8; 20]) -> Result<()
 
     */
 
+    is_checker_quit_sidechain()?;
+
     let witness = CheckerQuitSidechainWitness::from_raw(raw_witness).ok_or(Error::Encoding)?;
 
     let (config_input, checker_bond_input_lock_args, checker_bond_input, checker_info_input) = load_entities! {
@@ -51,11 +53,11 @@ pub fn checker_quit_sidechain(raw_witness: &[u8], signer: [u8; 20]) -> Result<()
         return Err(Error::SidechainConfigMismatch);
     }
     config_res.checker_total_count -= 1;
-
-    config_res.checker_bitmap = bit_map_remove(config_res.checker_bitmap, witness.checker_id)?;
+    config_res.checker_bitmap = bit_map_remove(config_res.checker_bitmap, witness.checker_id).ok_or(Error::SidechainConfigMismatch)?;
 
     let mut checker_bond_res_lock_args = checker_bond_input_lock_args.clone();
-    checker_bond_res_lock_args.chain_id_bitmap = bit_map_remove(checker_bond_res_lock_args.chain_id_bitmap, witness.chain_id)?;
+    checker_bond_res_lock_args.chain_id_bitmap =
+        bit_map_remove(checker_bond_res_lock_args.chain_id_bitmap, witness.chain_id).ok_or(Error::CheckerBondMismatch)?;
 
     has_sidechain_config_passed_update_interval(config_input, CONFIG_INPUT)?;
     if config_res.chain_id != witness.chain_id || config_res != config_output {
@@ -74,6 +76,33 @@ pub fn checker_quit_sidechain(raw_witness: &[u8], signer: [u8; 20]) -> Result<()
     {
         return Err(Error::CheckerInfoMismatch);
     }
+
+    Ok(())
+}
+
+fn is_checker_quit_sidechain() -> Result<(), Error> {
+    let global = check_global_cell()?;
+
+    let input_count = get_input_cell_count();
+    let output_count = get_output_cell_count();
+
+    if input_count != 4 || output_count != 3 {
+        return Err(Error::CellNumberMismatch);
+    }
+
+    check_cells! {
+        &global,
+        {
+            CodeCellData: CODE_INPUT,
+            SidechainConfigCellData: CONFIG_INPUT,
+            CheckerBondCellData: CHECKER_BOND_INPUT,
+            CheckerInfoCellData: CHECKER_INFO_INPUT,
+
+            CodeCellData: CODE_OUTPUT,
+            SidechainConfigCellData: CONFIG_OUTPUT,
+            CheckerBondCellData: CHECKER_BOND_OUTPUT,
+        },
+    };
 
     Ok(())
 }
