+++ axon-ckb-contracts/contracts/code-cell/src/checker_submit_task.rs
@@ -3,20 +3,22 @@ use ckb_std::ckb_constants::Source;
 use common_raw::{
     cell::{
         checker_info::{CheckerInfoCellData, CheckerInfoCellMode},
+        code::CodeCellData,
         sidechain_config::SidechainConfigCellData,
         task::{TaskCellData, TaskCellMode},
     },
     witness::checker_submit_task::CheckerSubmitTaskWitness,
+    FromRaw,
 };
 
-use crate::{common::*, error::Error};
+use crate::{cell::*, common::*, error::Error};
 
 const CHECKER_INFO_INPUT: CellOrigin = CellOrigin(1, Source::Input);
 const TASK_INPUT: CellOrigin = CellOrigin(2, Source::Input);
 
 const CHECKER_INFO_OUTPUT: CellOrigin = CellOrigin(1, Source::Output);
 
-pub fn checker_submit_task(witness: &CheckerSubmitTaskWitness, signer: [u8; 20]) -> Result<(), Error> {
+pub fn checker_submit_task(raw_witness: &[u8], signer: [u8; 20]) -> Result<(), Error> {
     /*
     CheckerSubmitTask,
 
@@ -28,6 +30,10 @@ pub fn checker_submit_task(witness: &CheckerSubmitTaskWitness, signer: [u8; 20])
     Task Cell                   ->          Null
 
     */
+
+    let witness = CheckerSubmitTaskWitness::from_raw(raw_witness).ok_or(Error::Encoding)?;
+
+    is_checker_submit_task(&witness)?;
 
     let config_dep = SidechainConfigCellData::load(CellOrigin(witness.sidechain_config_dep_index, Source::CellDep))?;
     let (checker_info_input, task_input) = load_entities! {
@@ -51,6 +57,43 @@ pub fn checker_submit_task(witness: &CheckerSubmitTaskWitness, signer: [u8; 20])
     if task_input.chain_id != witness.chain_id || task_input.mode != TaskCellMode::Task {
         return Err(Error::TaskMismatch);
     }
+
+    Ok(())
+}
+
+fn is_checker_submit_task(witness: &CheckerSubmitTaskWitness) -> Result<(), Error> {
+    /*
+    CheckerSubmitTask,
+
+    Dep:    0 Global Config Cell
+    Dep:    1 Sidechain Config Cell
+
+    Code Cell                   ->         Code Cell
+    Checker Info Cell           ->          Checker Info Cell
+    Task Cell                   ->          Null
+
+    */
+
+    let global = check_global_cell()?;
+
+    let input_count = get_input_cell_count();
+    let output_count = get_output_cell_count();
+
+    if input_count != 3 || output_count != 2 {
+        return Err(Error::CellNumberMismatch);
+    }
+
+    check_cells! {
+        &global,
+        {
+            SidechainConfigCellData: CellOrigin(witness.sidechain_config_dep_index, Source::CellDep),
+            CodeCellData: CellOrigin(0, Source::Input),
+            CheckerInfoCellData: CellOrigin(1, Source::Input),
+            TaskCellData: CellOrigin(2, Source::Input),
+            CodeCellData: CellOrigin(0, Source::Output),
+            CheckerInfoCellData: CellOrigin(1, Source::Output),
+        },
+    };
 
     Ok(())
 }
