+++ axon-ckb-contracts/contracts/sidechain-config-cell-typescript/src/error.rs
@@ -1,5 +1,4 @@
 use ckb_std::error::SysError;
-use common::error::CommonError;
 
 /// Error
 #[repr(i8)]
@@ -9,7 +8,7 @@ pub enum Error {
     ItemMissing,
     LengthNotEnough,
     Encoding,
-    MissingTypeScript,
+    CodeCellMissing,
 }
 
 impl From<SysError> for Error {
@@ -21,15 +20,6 @@ impl From<SysError> for Error {
             LengthNotEnough(_) => Self::LengthNotEnough,
             Encoding => Self::Encoding,
             Unknown(err_code) => panic!("unexpected sys error {}", err_code),
-        }
-    }
-}
-
-impl From<CommonError> for Error {
-    fn from(err: CommonError) -> Self {
-        match err {
-            CommonError::MissingTypeScript => Self::MissingTypeScript,
-            _ => Self::MissingTypeScript,
         }
     }
 }
