+++ axon-ckb-contracts/contracts/sidechain-fee-cell-lockscript/src/entry.rs
@@ -1,7 +1,7 @@
 use core::result::Result;
 
 use crate::error::Error;
-use common::pattern::check_code_cell;
+use common::check_code_cell;
 
 pub fn main() -> Result<(), Error> {
     /*
@@ -12,7 +12,7 @@ pub fn main() -> Result<(), Error> {
     3. CheckerTakeBeneficiary
     */
 
-    check_code_cell()?;
+    check_code_cell().ok_or(Error::CodeCellMissing)?;
 
     Ok(())
 }
