+++ axon-ckb-contracts/contracts/task-cell-typescript/src/entry.rs
@@ -2,7 +2,7 @@ use core::result::Result;
 
 use crate::error::Error;
 
-use common::pattern::check_code_cell;
+use common::check_code_cell;
 
 pub fn main() -> Result<(), Error> {
     /*
@@ -17,7 +17,7 @@ pub fn main() -> Result<(), Error> {
     7. CheckerSubmitChallenge
     */
 
-    check_code_cell()?;
+    check_code_cell().ok_or(Error::CodeCellMissing)?;
 
     Ok(())
 }
