+++ axon-ckb-contracts/contracts/task-cell-typescript/src/error.rs
@@ -1,5 +1,4 @@
 use ckb_std::error::SysError;
-use common::error::CommonError;
 
 /// Error
 #[repr(i8)]
@@ -8,6 +7,7 @@ pub enum Error {
     ItemMissing,
     LengthNotEnough,
     Encoding,
+    CodeCellMissing,
 }
 
 impl From<SysError> for Error {
@@ -20,11 +20,5 @@ impl From<SysError> for Error {
             Encoding => Self::Encoding,
             Unknown(err_code) => panic!("unexpected sys error {}", err_code),
         }
-    }
-}
-
-impl From<CommonError> for Error {
-    fn from(_: CommonError) -> Self {
-        unimplemented!()
     }
 }
